requirements.txt
asgiref==3.5.0
Django==4.0.3
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
djangorestframework==3.13.1
phonenumbers==8.12.44
python-decouple==3.6
pytz==2021.3
sqlparse==0.4.2
